feat: Reorganize DML and DDL visitors ready for CREATE/ALTER implementation
# Conflicts:
#	core/src/main/antlr4/com/databricks/labs/remorph/parsers/tsql/TSqlParser.g4
#	core/src/main/scala/com/databricks/labs/remorph/parsers/tsql/TSqlAstBuilder.scala
#	core/src/main/scala/com/databricks/labs/remorph/parsers/tsql/TSqlRelationBuilder.scala
#	core/src/main/scala/com/databricks/labs/remorph/parsers/tsql/relations.scala
