+++ core/src/main/scala/com/databricks/labs/remorph/parsers/tsql/TSqlAstBuilder.scala
@@ -1,5 +1,6 @@
 package com.databricks.labs.remorph.parsers.tsql
 
+import com.databricks.labs.remorph.parsers.{intermediate => ir}
 import com.databricks.labs.remorph.parsers.intermediate.LogicalPlan
 import com.databricks.labs.remorph.parsers.tsql.TSqlParser.DmlClauseContext
 import com.databricks.labs.remorph.parsers.{OptionAuto, OptionExpression, OptionOff, OptionOn, OptionString, intermediate => ir}
@@ -13,7 +14,8 @@ import scala.collection.JavaConverters.asScalaBufferConverter
 class TSqlAstBuilder extends TSqlParserBaseVisitor[ir.LogicalPlan] {
 
   private val relationBuilder = new TSqlRelationBuilder
-  private val optionBuilder = new OptionBuilder(new TSqlExpressionBuilder)
+  private val dmlBuilder = new TSqlDMLBuilder
+  private val ddlBuilder = new TSqlDDLBuilder
 
   override def visitTSqlFile(ctx: TSqlParser.TSqlFileContext): ir.LogicalPlan = {
     Option(ctx.batch()).map(_.accept(this)).getOrElse(ir.Batch(List()))
@@ -38,9 +40,9 @@ class TSqlAstBuilder extends TSqlParserBaseVisitor[ir.LogicalPlan] {
       case dml if dml.dmlClause() != null => dml.dmlClause().accept(this)
       case cfl if cfl.cflStatement() != null => cfl.cflStatement().accept(this)
       case another if another.anotherStatement() != null => another.anotherStatement().accept(this)
-      case ddl if ddl.ddlClause() != null => ddl.ddlClause().accept(this)
+      case ddl if ddl.ddlClause() != null => ddl.ddlClause().accept(ddlBuilder)
       case dbcc if dbcc.dbccClause() != null => dbcc.dbccClause().accept(this)
-      case backup if backup.backupStatement() != null => backup.backupStatement().accept(this)
+      case backup if backup.backupStatement() != null => backup.backupStatement().accept(ddlBuilder)
       case coaFunction if coaFunction.createOrAlterFunction() != null =>
         coaFunction.createOrAlterFunction().accept(this)
       case coaProcedure if coaProcedure.createOrAlterProcedure() != null =>
@@ -65,35 +67,19 @@ class TSqlAstBuilder extends TSqlParserBaseVisitor[ir.LogicalPlan] {
     }
   }
 
-  /**
-   * This is not actually implemented but was a quick way to exercise the genericOption builder before we had other
-   * syntax implemented to test it with.
-   *
-   * @param ctx
-   *   the parse tree
-   */
-  override def visitBackupStatement(ctx: TSqlParser.BackupStatementContext): ir.LogicalPlan = {
-    ctx.backupDatabase().accept(this)
-  }
-
-  override def visitBackupDatabase(ctx: TSqlParser.BackupDatabaseContext): ir.LogicalPlan = {
-    val database = ctx.id().getText
-    val opts = ctx.optionList()
-    val options = opts.asScala.flatMap(_.genericOption().asScala).toList.map(optionBuilder.buildOption)
-    val (disks, boolFlags, autoFlags, values) = options.foldLeft(
-      (List.empty[String], Map.empty[String, Boolean], List.empty[String], Map.empty[String, ir.Expression])) {
-      case ((disks, boolFlags, autoFlags, values), option) =>
-        option match {
-          case OptionString("DISK", value) =>
-            (value.stripPrefix("'").stripSuffix("'") :: disks, boolFlags, autoFlags, values)
-          case OptionOn(id) => (disks, boolFlags + (id -> true), autoFlags, values)
-          case OptionOff(id) => (disks, boolFlags + (id -> false), autoFlags, values)
-          case OptionAuto(id) => (disks, boolFlags, id :: autoFlags, values)
-          case OptionExpression(id, expr, _) => (disks, boolFlags, autoFlags, values + (id -> expr))
-          case _ => (disks, boolFlags, autoFlags, values)
-        }
+  override def visitDmlClause(ctx: TSqlParser.DmlClauseContext): ir.LogicalPlan = {
+    val dml = ctx match {
+      case dml if dml.selectStatement() != null =>
+        dml.selectStatement().accept(relationBuilder)
+      case _ =>
+        ctx.accept(dmlBuilder)
     }
-    // Default flags generally don't need to be specified as they are by definition, the default
-    BackupDatabase(database, disks, boolFlags, autoFlags, values)
+
+    Option(ctx.withExpression())
+      .map { withExpression =>
+        val ctes = withExpression.commonTableExpression().asScala.map(_.accept(relationBuilder))
+        ir.WithCTE(ctes, dml)
+      }
+      .getOrElse(dml)
   }
 }
